requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==36.0.1
decorator==5.0.5
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.47.0
django-crispy-forms==1.13.0
django-extensions==3.1.2
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.0.12
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Table==0.5.0
gcloud==0.18.3
googleapis-common-protos==1.54.0
gunicorn==20.1.0
httplib2==0.20.2
idna==2.10
import-ipynb==0.1.3
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-core==4.7.1
jwcrypto==1.0
MarkupSafe==1.1.1
mysqlclient==2.0.3
nbformat==5.1.3
numpy==1.20.2
oauth2client==4.1.3
oauthlib==3.1.1
pandas==1.2.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.18
protobuf==3.19.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
Pygments==2.8.1
PyJWT==2.3.0
pyparsing==3.0.6
Pyrebase4==4.5.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-jwt==3.3.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.8
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
traitlets==5.0.5
urllib3==1.26.4
virtualenv==20.4.3
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.3.0
wrapt==1.13.3
